Updated client-setting service
